pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sa-segformer-rts"
version = "0.1.0"
description = "SA-SegFormer training and evaluation code for retrogressive thaw slump segmentation."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "SA-SegFormer RTS contributors" }]
dependencies = [
  "albumentations>=1.3",
  "matplotlib>=3.7",
  "numpy>=1.23",
  "opencv-python>=4.8",
  "openpyxl>=3.1",
  "pandas>=2.0",
  "Pillow>=10.0",
  "segmentation-models-pytorch>=0.3",
  "tifffile>=2023.7",
  "timm>=0.9",
  "torch>=2.0",
  "torchvision>=0.15",
  "tqdm>=4.66",
]

[tool.setuptools.packages.find]
where = ["src"]
